+++ byway/src/main/webapp/index.js
@@ -1,10 +1,10 @@
 window.onload = function(){ 
   fetch("/login").then(response => response.json()).then((login) =>{
   if (status == false) {
-    let loginLink= document.createElement("a");
+    let loginLink = document.createElement("a");
     loginLink.id = "login-button";
     loginLink.href = login.url;
-    loginLink.innerText= "LOGIN";
+    loginLink.innerText = "LOGIN";
     let container = document.getElementById("login-link");
     container.href = login.url;
     container.append(loginLink);
